+++ task_1_5.py
@@ -1,4 +1,4 @@
-a = float(input("Enter triangle leg1:"))
-b = float(input("Enter triangle leg2:"))
+a = int(input("Enter triangle leg1:"))
+b = int(input("Enter triangle leg2:"))
 print("Hypotenuse length:", ((a**2+b**2)**0.5))
 print("Tringle area:", a*b/2)
